requirements.txt
asgiref==3.2.5
certifi==2020.4.5.1
chardet==3.0.4
Django==2.1.11
django-debug-toolbar==2.2
djangorestframework==3.11.0
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
